attempt 101 to fix comments

+++ src/main/java/com/codeup/dogdays/controllers/CommentController.java
@@ -54,7 +54,7 @@ public class CommentController {
         comment.setEvents(eventRepo.findOne(id));
 
         List<Comment> allComments = (List<Comment>)commentRepo.findAll();
-        int count = allComments.size();
+        long count = allComments.size();
